requirements.txt
fastapi
marker-pdf
pydantic
scalar_fastapi